requirements.txt
backports.entry-points-selectable==1.1.1
certifi==2022.6.15
charset-normalizer==2.1.1
distlib==0.3.4
filelock==3.6.0
idna==3.3
infoblox-netmri==3.8.0.0
pbr==5.8.0
platformdirs==2.5.2
python-dotenv==0.20.0
PyYAML==6.0
requests==2.28.1
six==1.16.0
stevedore==3.5.0
urllib3==1.26.12
virtualenv==20.14.1
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
virtualenvwrapper.django==0.4.1
